+++ back/config/db-config.js
@@ -1,5 +1,5 @@
 const { Sequelize } = require('sequelize');
-const { createUsersOnStart, createRolesOnStart } = require('../seeders/populateDB');
+// const { createUsersOnStart, createRolesOnStart } = require('../seeders/populateDB');
 
 const sequelize = new Sequelize(process.env.DB_DATABASE, process.env.DB_USER, process.env.DB_PASSWORD, {
   host: process.env.DB_HOST,
