+++ back/controllers/index.js
@@ -1,5 +1,7 @@
 const userController = require('./user.js');
+const loginController = require('./login.js');
 
 module.exports = {
-  userController
+  userController,
+  loginController
 };
